NewsFetcherAgent/requirements.txt
feedparser==6.0.11
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0